Don't suggest users add ~/Downloads to $PATH
IMHO this is terrible advice, because who knows what evil stuff will be in their `~/Downloads` directory, and now that's suddenly in their `$PATH`.  I've modified the suggestions to hopefully nudge people to install using the packaged versions and if they don't, just get them to execute using the absolute path to the downloaded file.