Doc scrub - section rewrite* (#1173)
Rewrote the "Subscribing to AWS Marketplace" section for accuracy. Previously, the section addressed only AWS customers and used the term "VM-series," which is specific to Palo Alto and not a general term for a firewall instance. Rewrote the section to make it clear that only AWS customers need to subscribe to a firewall instance and clarifying that a VM series for Palo Alto = CloudGuard for Check Point = "Fortigate" for Fortinet.