Removing broken links
Removed the broken links to the Aviatrix Terraform Tutorial and the second document, as those no longer exist in the repository. The old copy of the tutorial was too out-of-date to be useful. I substituted a link to the general Terraform Registry after discussing with Anthony Lee.

I also changed "Setup" to "Setting up" and made a few more minor formatting fixes.